requirements.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile requirements.in
#
certifi==2022.6.15
    # via requests
charset-normalizer==2.1.0
    # via requests
idna==3.3
    # via requests
pillow==9.2.0
    # via -r requirements.in
pydantic==1.9.1
    # via -r requirements.in
requests==2.28.1
    # via -r requirements.in
typing-extensions==4.3.0
    # via pydantic
urllib3==1.26.11
    # via requests
